django_projects/RecipeRealm/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.1
django-ckeditor==6.5.1
django-js-asset==2.0.0
mysqlclient==2.1.1
Pillow==9.5.0
python-dotenv==1.0.0
sqlparse==0.4.4
typing-extensions==4.6.2
